Refactoring done
